Move domain decomposer and metadata
