+++ include/mgard-x/DomainDecomposer/CMakeLists.txt
@@ -1,0 +1,4 @@
+list(APPEND MGARD_X_HEADER
+    ${CMAKE_CURRENT_SOURCE_DIR}/DomainDecomposer.hpp
+    )
+set(MGARD_X_HEADER ${MGARD_X_HEADER} PARENT_SCOPE)
